Integrate bedfile database into existing database
Added bedfile databse and created a foregin key relationship to
the genes database
