--- test_expected_4562.bed
@@ -1,6 +1,0 @@
-chromosome	start	end	name	score	strand
-chr17	18950137	18950295	NC_000017.10_GRAP_exon_1	0	-
-chr17	18945095	18945242	NC_000017.10_GRAP_exon_2	0	-
-chr17	18939246	18939418	NC_000017.10_GRAP_exon_3	0	-
-chr17	18927503	18927721	NC_000017.10_GRAP_exon_4	0	-
-chr17	18923944	18925482	NC_000017.10_GRAP_exon_5	0	-
